refactor: Wish 패키지 분리
