+++ src/main/java/com/example/travelday/domain/wish/entity/Wish.java
@@ -1,6 +1,7 @@
-package com.example.travelday.domain.travelroom.entity;
+package com.example.travelday.domain.wish.entity;
 
-import com.example.travelday.domain.travelroom.dto.request.WishReqDto;
+import com.example.travelday.domain.wish.dto.request.WishReqDto;
+import com.example.travelday.domain.travelroom.entity.TravelRoom;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Builder;
